Refactor sync tests and samples
