+++ source/sdk/react-native/quick-start.txt
@@ -249,9 +249,8 @@ To use App Services features, such as authentication and sync, you must first ac
 your App Services App using your App ID. You can :ref:`find your App ID <find-your-app-id>`
 in the App Services UI.
 
-.. literalinclude:: /examples/generated/react-native/ts/quickstart-sync.test.snippet.configure-expose-realm.tsx
+.. literalinclude:: /examples/generated/react-native/ts/quickstart-sync-app-provider.test.snippet.app-provider.tsx
   :language: javascript
-  :emphasize-lines: 4, 14
 
 Authenticate a User with <UserProvider>
 ```````````````````````````````````````
@@ -267,9 +266,8 @@ When :ref:`anonymous authentication <anonymous-authentication>` is enabled,
 users can immediately log into your app without providing any identifying
 information:
 
-.. literalinclude:: /examples/generated/react-native/ts/quickstart-sync.test.snippet.configure-expose-realm.tsx
+.. literalinclude:: /examples/generated/react-native/ts/quickstart-sync-user-provider.test.snippet.user-provider.tsx
   :language: javascript
-  :emphasize-lines: 5, 13
 
 Configure a Synced Realm with <RealmProvider>
 `````````````````````````````````````````````
@@ -299,7 +297,7 @@ You can add subscriptions in your components or set up
 
 .. literalinclude:: /examples/generated/react-native/ts/quickstart-sync.test.snippet.configure-expose-realm.tsx
   :language: javascript
-  :emphasize-lines: 6-12
+  :emphasize-lines: 6-19
 
 The syntax to create, read, update, and delete objects in a synced realm is
 identical to the syntax for non-synced realms. While you work with local data, a
@@ -311,7 +309,7 @@ Next: Check out the Template Apps and Tutorial
 Check out the :ref:`template app <template-apps>` to experiment with another
 way to get started with the Realm React Native SDK. 
 The React Native SDK template app is a customizable application built on 
-Realm Database (local-only or Flexible Sync).
+a Flexible Sync Realm database.
 
 Alternatively, if you are interested in a guided experience, you can read our
 :ref:`Realm React Native SDK tutorial <react-native-tutorial>` that expands on
